html and css layout
The thing that I explored most was the location of each division. It
was quite challenging because the size of the box kept change when I
add some padding and margin options. Also situating the main image was
somewhat hard cause I couldn’t quite get the size of the image that I
wanted. However, I’m glad that I finally figured out.